Added batch process prioritization [#2455]
 * Each processor checks the comic's flags before processing.
 * Any comic that does not pass the test is not processed.

The item processor for each batch process will check to see if
the comic is in a state that is allowed for that process to run.

+++ comixed-batch/src/main/java/org/comixedproject/batch/comicbooks/processors/PurgeMarkedComicsProcessor.java
@@ -48,6 +48,10 @@ public class PurgeMarkedComicsProcessor implements ItemProcessor<ComicBook, Comi
 
   @Override
   public ComicBook process(final ComicBook comicBook) throws Exception {
+    if (comicBook.isFileContentsLoaded() == false) {
+      log.debug("Comic not ready for purging: id={}", comicBook.getComicBookId());
+      return null;
+    }
     try {
       log.debug("Removing comic book from all reading lists: id={}", comicBook.getComicBookId());
       this.readingListService.deleteEntriesForComicBook(comicBook);
